12/20/2018 Update
Error with inputting A as the second nucleotide base is fixed, and now it takes input from the console, rather than just putting in the different nucleotide bases directly into the source code. Another change is that it loops until you specify no, meaning that you can put multiple nucleotide triplets in only one run of the program.